+++ src/apps/sequencer/engine/TrackEngine.h
@@ -88,6 +88,9 @@ public:
 
     virtual float sequenceProgress() const { return -1.f; }
 
+    // For look-ahead: set per-track first-step-after-start
+    virtual void setFirstStepAfterStart(bool value) {}
+
     // helpers
 
     bool isSelected() const { return _model.project().selectedTrackIndex() == _track.trackIndex(); }